requirements.txt
blinker==1.9.0
click==8.2.1
config==0.5.1
flask==3.1.1
flask-jwt-extended==4.7.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
pyjwt==2.10.1
repoze-lru==0.7
routes==2.5.1
six==1.17.0
sqlalchemy==2.0.42
typing-extensions==4.14.1
werkzeug==3.1.3
wtforms==3.2.1
